Fix sound bug
When user clicks the wrong button, the sound of mistake starts, but
after the initial button sound has played. This way, the user can not
know if he made a mistake before the button was clicked,just by
holding the mouse down and leaving the button area to release the
mouse click. Now the user is prevented from one way of cheating the game